Simplified transaction download login (now downloads year-to-date). Removed redundant check of "Continue" button presence. Removed code never used.
